Change name
There is no reason for an open source project with so much community interest to have response times to issue and pull requests measured in weeks. I'm creating a fork of mocha that I hope can be opened to management by a (small) group. This will be an interesting experiment.